backend/requirements.txt
asgiref 
Django 
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg==3.2.3
sqlparse 
typing_extensions 
